backend/requirements.txt
fastapi
uvicorn
python-multipart
websockets
torch==2.1.1+cpu
torchvision==0.16.1+cpu
numpy==1.26.2
opencv-python==4.8.1.78
ultralytics==8.0.200
mmaction2==1.2.0
mmcv==2.1.0
mmengine==0.10.1